Add Option support
The flag package doesn't support options as supported by the getopt_long
function of the GNU libc library. So we want our own implementation. From my
point of view the flag API has a too wide interface. Particularly annoying is
FlagSet which is basically a list of flags.

So we implemented our own version using the option term used by the GNU C
library. The library defines an Option structure that can be used by the client
to define custom option. A single function is provided for default use cases
like boolean or string flags.

The function UsageOption prints usage information for the options and
ParseOptions parses arguments using a slice of options.