Reimplemnted UsageOptions
The UsageOptions have been reimplemented. The description of a flag has now been
split in two part. The first line explains the usage of the flags and the second
is a description that is been reformatted to fit an 80 column line width. This
required the formatText function to be implemented. It uses a lexer to
distinguish normal words, verbatim sections and paragraphs.